Added TraceHelper and AceUtils classes. Corrected faults in EsInstrumenter. Current beta for Aurelia merge. Eval like fucntionality still pending
