--- src/ApplicationState.js
@@ -1,8 +1,0 @@
-export class ApplicationState {  
-  constructor () {
-    this.loggedInUser = null;
-    this.configuration  = {
-            tracer : undefined
-        };
-  }
-}
