+++ src/jsEditor/js-editor.html
@@ -1,4 +1,3 @@
 <template bindable="editor">
     <div id="editorDiv" style="height: 500px"></div>
-    <div id="shadowEditorDiv" style="height: 0px"></div>
 </template>
