+++ src/pastebin/pastebin.js
@@ -4,17 +4,20 @@ import {Router} from 'aurelia-router';
 import {JsEditor} from '../jsEditor/js-editor';
 import {JsGutter} from '../jsGutter/js-gutter';
 import {ConsoleWindow} from '../consoleWindow/console-window';
+import {JsSandBox} from '../traceService/js-sandbox';
 
-@inject(EventAggregator, Router, JsEditor, JsGutter, ConsoleWindow)
+@inject(EventAggregator, Router, JsEditor, JsGutter, ConsoleWindow, JsSandBox)
 export class Pastebin {
 
-  constructor(eventAggregator, router, jsEditor, jsGutter, consoleWindow) {
+  constructor(eventAggregator, router, jsEditor, jsGutter, consoleWindow, jsSandBox) {
     this.eventAggregator = eventAggregator;
     this.router = router;
     this.heading = 'Pastebin';
     this.jsEditor = jsEditor;
     this.jsGutter = jsGutter;
     this.consoleWindow = consoleWindow;
+    
+    this.jsSandBox = jsSandBox;
   }
 
   activate(params) {
@@ -37,6 +40,7 @@ export class Pastebin {
     this.jsEditor.attached();
     this.jsGutter.attached();
     this.consoleWindow.attached();
+    this.jsSandBox.attached();
   }
 
   subscribe() {
