+++ src/traceService/js-sandbox.html
@@ -1,0 +1,3 @@
+<template bindable="sandbox">
+    <div id="sandBoxDiv" style="height: 1000px"></div>
+</template>
