+++ src/traceService/traceHelper.js
@@ -1,0 +1,64 @@
+
+export class TraceHelper {
+    constructor(executionTrace){
+        this.executionTrace = executionTrace;
+    }
+    // example of how to use the trace resulting data structure
+    visualizeExecutionTrace(executionTrace){
+        var i, entry;
+        var stackText= "";
+
+        for (i = 0; i < executionTrace.length; i += 1) {
+            entry = executionTrace[i];
+            stackText += i + " -- " + JSON.stringify(entry) + "<br> ";
+           
+        }
+       
+    	
+    	return stackText;  
+    
+    }
+       // example of how to use the trace resulting data structure
+    visualize(stackTrace){
+        var i, entry, name, index;
+        var stackText= "";
+    	var repeat = 0;
+    	var previousCall = "";
+    	var previousIndex = -1;
+        for (i = 0; i < stackTrace.length; i += 1) {
+            entry = stackTrace[i];
+            if(entry){
+                name = entry.text;
+        		index = entry.index;
+        		 if(previousCall !== name){				 
+        			 if(repeat > 0){
+        				 stackText += previousIndex + " -- " + previousCall + "( + "+ repeat +" times) <br>";
+        				 repeat = 0;
+        			 }else{
+        				 if(previousIndex > -1){
+        					 stackText += previousIndex + " -- " + previousCall + "<br> ";
+        				 }
+        				 
+        			 }
+        			 previousCall = name;
+        			 previousIndex = index;
+        		 }else{
+        			 repeat = repeat + 1; 
+        		 }
+            }
+    		
+        }
+    	if(repeat > 0){
+    		stackText +=  previousIndex + " -- " + previousCall + "( + "+ repeat +" times )";
+    		repeat = 0;
+    	}else{
+    		if(previousIndex > -1){
+    			stackText += previousIndex + " -- " + previousCall ;
+    		}					 
+    	}
+    	
+    	return stackText;  
+    
+    }
+    
+}
