+++ src/traceService/traceService.js
@@ -1,42 +1,29 @@
 import {inject} from 'aurelia-framework';
 import {EventAggregator} from 'aurelia-event-aggregator';
-import {ApplicationState} from '../ApplicationState';
 
 //local imports
 import {EsTracer} from './estracer';
 
-@inject(EventAggregator, ApplicationState)
+@inject(EventAggregator)
 export class TraceService {
 
-    constructor(eventAggregator, appState) {
+    constructor(eventAggregator) {
       // inject code in the service. Dependencies failure happens here
       this.eventAggregator = eventAggregator;
-      this.appState = appState;
+
       this.events = {
         onTraceServiceStart: {  event :'onTraceServiceStart', description : 'Building Tracer...' },
         onTraceServiceRunning: {  event :'onTraceServiceRunning', description : 'Tracing...' },
-        onTraceServiceEnd :{  event :'onTraceServiceEnd', description : 'Trace build succesfully.' },
+        onTraceServiceEnd :{  event :'onTraceServiceEnd', description : 'Trace build successfully.' },
         onTraceServiceBusy : {  event :'onTraceServiceBusy', description : 'A previous trace is running.' },
         onTraceServiceException: {  event :'onTraceServiceException', description : 'Trace ended due to an exception.' }
       };
       
       this.timeLimit = 3000; //default timeout
       
-      //singleton?
-      if(this.appState && this.appState.configuration.tracer){
-          this.estracer = this.appState.configuration.tracer;
-          
-            console.log("Old Instance");
-      }else{
-        this.estracer = new EsTracer(this.events);
-        console.log("New Instance");
-        if(this.appState){
-            console.log("New Instance saved");
-            this.appState.configuration.tracer = this.estracer;
-        }
-      }
-      
-     
+
+      this.estracer = new EsTracer(this.events);
+
     }
 // exec-viz suscribe to the onTraceChanged event
 //   onTraceChanged(result){
@@ -58,7 +45,7 @@ export class TraceService {
        this.timeLimit = timeLimit; 
     }
     cancelTrace(){
-        this.estracer.ISCANCELLED = true;
+        window.ISCANCELLED = true;
     }
     
     getTrace(code, publisher) { //optional publisher
@@ -74,122 +61,8 @@ export class TraceService {
         
     }
     
-
-    // example using ACE Annotations
-    showTraceAnnotations(aceEditor) {
-            var annotations = aceEditor.getSession().getAnnotations();	
-            annotations= annotations.concat(traceAnnotations);
-    		aceEditor.getSession().setAnnotations(annotations);
-    }
-    
-    // example of how to use the trace resulting data structure
-    visualize(stackTrace){
-        var i, entry, name, index;
-        var stackText= "";
-    	var repeat = 0;
-    	var previousCall = "";
-    	var previousIndex = -1;
-        for (i = 0; i < stackTrace.length; i += 1) {
-            entry = stackTrace[i];
-            if(entry){
-                name = entry.text;
-        		index = entry.index;
-        		 if(previousCall !== name){				 
-        			 if(repeat > 0){
-        				 stackText += previousIndex + " -- " + previousCall + "( + "+ repeat +" times) <br>";
-        				 repeat = 0;
-        			 }else{
-        				 if(previousIndex > -1){
-        					 stackText += previousIndex + " -- " + previousCall + "<br> ";
-        				 }
-        				 
-        			 }
-        			 previousCall = name;
-        			 previousIndex = index;
-        		 }else{
-        			 repeat = repeat + 1; 
-        		 }
-            }
-    		
-        }
-    	if(repeat > 0){
-    		stackText +=  previousIndex + " -- " + previousCall + "( + "+ repeat +" times )";
-    		repeat = 0;
-    	}else{
-    		if(previousIndex > -1){
-    			stackText += previousIndex + " -- " + previousCall ;
-    		}					 
-    	}
-    	
-    	return stackText;  
-    
-    }
-     // example of how to use the trace resulting data structure
-  visualizeExecutionTrace(executionTrace){
-        var i, entry;
-        var stackText= "";
-
-        for (i = 0; i < executionTrace.length; i += 1) {
-            entry = executionTrace[i];
-            stackText += i + " -- " + JSON.stringify(entry) + "<br> ";
-           
-        }
-       
-    	
-    	return stackText;  
-    
-    }
-    
-    isPositioninRange(position, inRange){
-        
-        var matchesInOneLine = (
-                position.row == inRange.start.row 
-                && inRange.start.row  == inRange.end.row
-                && position.column >= inRange.start.column
-                && position.column <= inRange.end.column
-            );
-            
-        if(matchesInOneLine){
-            return true;
-        }
-            
-        var matchesStart = (
-                position.row == inRange.start.row 
-                && inRange.start.row  < inRange.end.row
-                && position.column >= inRange.start.column
-            );
-           
-        if(matchesStart){
-            return true;
-        }
-        
-        var matchesEnd = (
-                position.row == inRange.end.row
-                && inRange.start.row  < inRange.end.row
-                && position.column <= inRange.end.column
-            );
-
-        return matchesEnd;
-
-    }
-    
-    isRangeInRange(isRange, inRange){
-        return (
-                (isRange.start.row >= inRange.start.row && isRange.start.column >= inRange.start.column)
-    			 &&
-    			(isRange.end.row <= inRange.end.row && isRange.end.column <= inRange.end.column)
-    			);
-    }
-    
-    isRangeInRangeStrict(isRange, inRange){
-        return (
-                (isRange.start.row >= inRange.start.row && isRange.start.column > inRange.start.column)
-    			 &&
-    			(isRange.end.row <= inRange.end.row && isRange.end.column < inRange.end.column)
-    			);
-    }
-    
     subscribe() {
+    // move this logic to the gutter
     // let ea = this.eventAggregator;
     // let session = this.session;
     
@@ -209,6 +82,7 @@ export class TraceService {
     //   }
     // });
   }
+  
   publish(event, payload){
      let ea = this.eventAggregator;
      if(ea){
